implemented Memory Bank Controller 1

+++ CM7/Core/Inc/stm32h7_display.h
@@ -9,7 +9,7 @@
 #define INC_STM32H7_DISPLAY_H_
 
 void stm32h7_displaySetPalette();
-void displayFrameBuffer(uint32_t* gb_frame, uint8_t scaleAmount);
+void displayFrameBuffer(uint8_t* gb_frame, uint8_t scaleAmount);
 void displayDoubleBuffer();
 void CopyBuffer(uint32_t *pSrc,
                            uint32_t *pDst,
